Add a toggle component and remove it as input component
